requirements.txt
aiohttp==3.12.14
ics==0.7.2
interactions.py==5.12.0
python-dotenv==1.0.1